beta version: fix addBook, editBook

+++ backend/server.js
@@ -2,14 +2,19 @@ import express from 'express'
 import cors from 'cors'
 import * as db_request from './queries.js'
 import { server_health } from './monitoring.js';
+import { fileURLToPath } from 'url';
 import fs from 'fs'
 import path from 'path'
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 
 const PORT = 3000;
 const app = express();
 
 const logFilePath = path.join(__dirname, 'log')
-fs.rmSync(logFilePath)
+if (fs.existsSync(logFilePath))
+  fs.rmSync(logFilePath);
 
 export function log(message) {
   const timestamp = new Date().toISOString();
@@ -30,7 +35,7 @@ app.get('/api/checklogin', async (req, res) => {
     const userLogin = req.query.login;
     const result = await db_request.checkLogin(userLogin);
     res.json({ result: result });
-    log('checkLogin success.')
+    log(`checkLogin success:\tLogin:\t{${userLogin}}`)
   } catch (error) {
     res.status(500).json({ error: error.message });
     log(`checkLogin failed:\t${error.message}`)
@@ -39,12 +44,11 @@ app.get('/api/checklogin', async (req, res) => {
 
 app.get('/api/checkuser', async (req, res) => {
   try {
-    console.log("checkUser")
     const userLogin = req.query.login;
     const userPassword = req.query.password;
     const result = await db_request.checkUser(userLogin, userPassword);
     res.json({ result: result });
-    log('checkUser success')
+    log(`checkUser success:\tUser:\t{${userLogin}, ${userPassword}}`)
   } catch (error) {
     res.status(500).json({ error: error.message });
     log(`checkUser failed:\t${error.message}`)
@@ -57,7 +61,7 @@ app.get('/api/createuser', async (req, res) => {
     const userPassword = req.query.password;
     const result = await db_request.createUser(userLogin, userPassword);
     res.json({ result: result });
-    log(`createUser success`)
+    log(`createUser success:\tUser:\t{${userLogin}, ${userPassword}}`)
   } catch (error) {
     res.status(500).json({ error: error.message });
     log(`createUser failed:\t${error.message}`)
@@ -71,7 +75,7 @@ app.post('/api/borrowbook', async (req, res) => {
     const book_id = borrow_data.book_id;
     const result = await db_request.borrowBook(user_name, book_id);
     res.json({ result: result });
-    log(`borrowBook success`)
+    log(`borrowBook success:\tBook:\t{${user_name}, ${book_id}}`)
   } catch (error) {
     res.status(500).json({ error: error.message });
     log(`borrowBook failed:\t${error.message}`)
@@ -83,7 +87,7 @@ app.post('/api/returnbook', async (req, res) => {
     const book_id = req.body.book_id;
     const result = await db_request.returnBook(book_id);
     res.json({ result: result });
-    log(`returnBook success`)
+    log(`returnBook success:\tBook:\t{${book_id}}`)
   } catch (error) {
     res.status(500).json({ error: error.message });
     log(`returnBook failed:\t${error.message}`)
@@ -95,7 +99,7 @@ app.get('/api/getbooks', async (req, res) => {
     const user_name = req.query.username;
     const result = await db_request.getBooks(user_name);
     res.json(result);
-    log(`getBooks success`)
+    log(`getBooks success:\tBook:\t{${user_name}}`)
   } catch (error) {
     res.status(500).json({ error: error.message });
     log(`getBooks failed:\t${error.message}`)
@@ -104,22 +108,22 @@ app.get('/api/getbooks', async (req, res) => {
 
 app.post('/api/addbook', async (req, res) => {
   try {
-    const book = req.body
+    const book = req.body.book
     const result = await db_request.addBook(book)
     res.json(result)
-    log(`addBook success`)
+    log(`addBook success:\tBook:\t{${book}}`)
   } catch (error) {
     res.status(500).json({ error: error.message });
     log(`addBook failed:\t${error.message}`)
   }
 })
 
-app.get('/api/deletebook', async (req, res) => {
+app.post('/api/deletebook', async (req, res) => {
   try {
-    const book_id = req.query.book_id
-    const result = await db_request.deleteBook(book_id)
+    const book_id = req.body.book_id
+    const result = await db_request.deleteBook(Number(book_id))
     res.json(result)
-    log(`deleteBook success`)
+    log(`deleteBook success:\tBook:\t{${book_id}}`)
   } catch (error) {
     res.status(500).json({ error: error.message });
     log(`deleteBook failed:\t${error.message}`)
@@ -130,7 +134,10 @@ app.get('/api/getusers', async (req, res) => {
   try {
     const result = await db_request.getUsers()
     res.json(result)
-    log(`getUsers success`)
+    const printUser = (user) => {
+      return `${user.user_id}, ${user.user_name}, ${user.user_password}, ${user.user_role}`
+    }
+    log(`getUsers success:\tUsers:\t[${result.map(user => `{${printUser(user)}}`)}]`)
   } catch (error) {
     res.status(500).json({ error: error.message })
     log(`getUsers failed:\t${error.message}`)
@@ -139,10 +146,10 @@ app.get('/api/getusers', async (req, res) => {
 
 app.post('/api/deleteuser', async (req, res) => {
   try {
-    const user_name = req.body.user_name;
+    const user_name = req.body.user_name
     const result = await db_request.deleteUser(user_name)
     res.json(result)
-    log(`deleteUser success`)
+    log(`deleteUser success:\tUser:\t{${user_name}}`)
   } catch (error) {
     res.status(500).json({ error: error.message })
     log(`deleteUser failed:\t${error.message}`)
@@ -153,10 +160,24 @@ app.get('/api/getloans', async (req, res) => {
   try {
     const result = await db_request.getLoans()
     res.json(result)
-    log(`getLoans success`)
+    log(`getLoans success:\tLoans:\t[${result}]`)
   } catch (error) {
     res.status(500).json({ error: error.message })
     log(`getLoans failed:\t${error.message}`)
+  }
+})
+
+app.post('/api/editbook', async (req, res) => {
+  try {
+    const book = req.body.book
+    const result = await db_request.editBook(book)
+    if (result[0].edit_book == false)
+      throw new Error("Failed to edit book")
+    res.json(result)
+    log(`editBook success:\tBook:\t[${book}]`)
+  } catch (error) {
+    res.status(500).json({ error: error.message })
+    log(`editBook failed:\t${error.message}`)
   }
 })
